requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
celery==5.3.6
redis==5.0.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
psycopg2-binary==2.9.11
